requirements.txt
transformers==4.41.1
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2
sentencepiece==0.2.0
Pillow==10.3.0
